pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
include = ["src/**"]

[tool.hatch.build.targets.wheel.force-include]
"credentials.json" = "credentials.json"
".env" = ".env"

[project]
name = "expense-tracker"
version = "0.1.0"
description = "Expense Tracker"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ruff",
    "python-dotenv",
    "google-api-python-client",
    "google-auth-httplib2",
    "google-auth-oauthlib",
    "bs4",
]

[project.optional-dependencies]
dev = ["pytest>=6.2.4"]

[project.scripts]
etrack = "src.cli.command:main"
gmailsync = "src.api.gmail:main"
